+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.1.0
+- [#188](https://github.com/cohere-ai/cohere-python/pull/188)
+  - Add `stream` parameter to chat, and relevant return object.
+- [#169](https://github.com/cohere-ai/cohere-python/pull/169)
+  - Add `stream` parameter to generate, and relevant return object.
+  - Add example notebook for streaming.
+
 ## 4.0.6
 - [#187](https://github.com/cohere-ai/cohere-python/pull/187)
   - Refactor feedback to be generate specific